src/requirements.txt
functions-framework==3.8.2
google-auth==2.38.0
google-cloud-storage==3.1.0
python-docx==1.1.2
qdrant-client==1.13.3
openai==1.70.0
pydantic-settings==2.8.1
redis==5.2.1
langfuse==2.60.2
